pipeline/requirements.txt
scrapy
user-agent
scrapy-user-agents
boto3
python-dotenv
apache-airflow
delta-spark 
psycopg2-binary 
supabase
fuzzywuzzy
RapidFuzz
feedparser
newspaper3k
PyYAML
lxml_html_clean
nanoid
pymilvus
sentence-transformers
numpy 
hf_xet
apache-airflow-providers-apache-spark
apache-airflow-providers-common-compat==1.2.2
pyspark==3.5.5
delta-spark==3.3.0

